added permissions to api-views-cards.py

+++ api/views/card.py
@@ -84,25 +84,34 @@ class SubCardViewSet(ViewSet):
 
     def create(self, request, card_id):
         card_instance = get_object_or_404(CardModel, id=card_id)
-        srz_data = SubCardSerializer(data=request.data)
-        if srz_data.is_valid():
-            srz_data.object.card = card_instance
-            srz_data.save()
-            return Response(srz_data.data, status=status.HTTP_201_CREATED)
-        return Response(srz_data.errors, status=status.HTTP_400_BAD_REQUEST)
-
+        if card_instance.cmembershipmodel_set.filter(to_user=request.user).exists():
+            srz_data = SubCardSerializer(data=request.data)
+            if srz_data.is_valid():
+                srz_data.object.card = card_instance
+                srz_data.save()
+                return Response(srz_data.data, status=status.HTTP_201_CREATED)
+            return Response(srz_data.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+           Response('permission denied, you are not a member of this card')
+           
     def partial_update(self, request, pk=None):
         subcard = get_object_or_404(self.queryset, pk=pk)
-        srz_data = SubCardSerializer(instance=subcard, data=request.POST, partial=True)
-        if srz_data.is_valid():
-            srz_data.save()
-            return Response(srz_data.data, status=status.HTTP_200_OK)
-        return Response(srz_data.errors, status=status.HTTP_400_BAD_REQUEST)
+        if subcard.card.cmembershipmodel_set.filter(to_user=request.user).exists():
+            srz_data = SubCardSerializer(instance=subcard, data=request.POST, partial=True)
+            if srz_data.is_valid():
+                srz_data.save()
+                return Response(srz_data.data, status=status.HTTP_200_OK)
+            return Response(srz_data.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+           Response('permission denied, you are not a member of this card')
 
     def destroy(self, request, pk):
         subcard = get_object_or_404(self.queryset, pk=pk)
-        subcard.delete()
-        return Response({'message': 'subcard deleted'}, status=status.HTTP_200_OK)
+        if subcard.card.cmembershipmodel_set.filter(to_user=request.user).exists():
+            subcard.delete()
+            return Response({'message': 'subcard deleted'}, status=status.HTTP_200_OK)
+        else:
+           Response('permission denied, you are not a member of this card')
 
 
 class ListViewSet(ViewSet):
@@ -110,48 +119,69 @@ class ListViewSet(ViewSet):
 
     def create(self, request, board_id):
         board_instance = get_object_or_404(BoardModel, id=board_id)
-        srz_data = ListSerializer(data=request.data)
-        if srz_data.is_valid():
-            srz_data.object.board = board_instance
-            srz_data.save()
-            return Response(srz_data.data, status=status.HTTP_201_CREATED)
-        return Response(srz_data.errors, status=status.HTTP_400_BAD_REQUEST)
+        if board_instance.owner == request.user:
+            srz_data = ListSerializer(data=request.data)
+            if srz_data.is_valid():
+                srz_data.object.board = board_instance
+                srz_data.save()
+                return Response(srz_data.data, status=status.HTTP_201_CREATED)
+            return Response(srz_data.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            Response('permission denied, you are not board owner')
+
 
     def partial_update(self, request, pk=None):
         list = get_object_or_404(self.queryset, pk=pk)
-        srz_data = ListSerializer(instance=list, data=request.POST, partial=True)
-        if srz_data.is_valid():
-            srz_data.save()
-            return Response(srz_data.data, status=status.HTTP_200_OK)
-        return Response(srz_data.errors, status=status.HTTP_400_BAD_REQUEST)
+        if list.board.owner == request.user:
+            srz_data = ListSerializer(instance=list, data=request.POST, partial=True)
+            if srz_data.is_valid():
+                srz_data.save()
+                return Response(srz_data.data, status=status.HTTP_200_OK)
+            return Response(srz_data.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            Response('permission denied, you are not board owner')
 
     def destroy(self, request, pk):
         list = get_object_or_404(self.queryset, pk=pk)
-        list.delete()
-        return Response({'message': 'list deleted'}, status=status.HTTP_200_OK)
-
+        if list.board.owner == request.user:
+            list.delete()
+            return Response({'message': 'list deleted'}, status=status.HTTP_200_OK)
+        else:
+            Response('permission denied, you are not board owner')
 
 class LabelViewSet(ViewSet):
     queryset = LabelModel.objects.all()
 
     def create(self, request, card_id):
         card_instance = get_object_or_404(CardModel, id=card_id)
-        srz_data = LabelSerializer(data=request.data)
-        if srz_data.is_valid():
-            srz_data.object.card = card_instance
-            srz_data.save()
-            return Response(srz_data.data, status=status.HTTP_201_CREATED)
-        return Response(srz_data.errors, status=status.HTTP_400_BAD_REQUEST)
+        if card_instance.cmembershipmodel_set.filter(to_user=request.user).exists()\
+            or card_instance.list.board.owner == request.user:
+            srz_data = LabelSerializer(data=request.data)
+            if srz_data.is_valid():
+                srz_data.object.card = card_instance
+                srz_data.save()
+                return Response(srz_data.data, status=status.HTTP_201_CREATED)
+            return Response(srz_data.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            Response('create label permission denied')
 
     def partial_update(self, request, pk=None):
         label = get_object_or_404(self.queryset, pk=pk)
-        srz_data = LabelSerializer(instance=label, data=request.POST, partial=True)
-        if srz_data.is_valid():
-            srz_data.save()
-            return Response(srz_data.data, status=status.HTTP_200_OK)
-        return Response(srz_data.errors, status=status.HTTP_400_BAD_REQUEST)
+        if label.card.cmembershipmodel_set.filter(to_user=request.user).exists()\
+            or label.card.list.board.owner == request.user:
+            srz_data = LabelSerializer(instance=label, data=request.POST, partial=True)
+            if srz_data.is_valid():
+                srz_data.save()
+                return Response(srz_data.data, status=status.HTTP_200_OK)
+            return Response(srz_data.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            Response('edit label permission denied')
 
     def destroy(self, request, pk):
         label = get_object_or_404(self.queryset, pk=pk)
-        label.delete()
-        return Response({'message': 'label deleted'}, status=status.HTTP_200_OK)
+        if label.card.cmembershipmodel_set.filter(to_user=request.user).exists()\
+        or label.card.list.board.owner == request.user:
+            label.delete()
+            return Response({'message': 'label deleted'}, status=status.HTTP_200_OK)
+        else:
+            Response('delete label permission denied')
